# Distribute `select count(*)` queries