--- a/conf/scylla.yaml
+++ b/conf/scylla.yaml
@@ -524,3 +524,7 @@ murmur3_partitioner_ignore_msb_bits: 12
 # Use a new optimized algorithm for performing reversed reads.
 # Set to `false` to fall-back to the old algorithm.
 # enable_optimized_reversed_reads: true
+
+# Use on a new, parallel algorithm for performing aggregate queries.
+# Set to `false` to fall-back to the old algorithm.
+# enable_parallelized_aggregation: true