--- a/cql3/query_processor.cc
+++ b/cql3/query_processor.cc
@@ -58,9 +58,10 @@ public:
     }
 };
 
-query_processor::query_processor(service::storage_proxy& proxy, data_dictionary::database db, service::migration_notifier& mn, service::migration_manager& mm, query_processor::memory_config mcfg, cql_config& cql_cfg)
+query_processor::query_processor(service::storage_proxy& proxy, service::forward_service& forwarder, data_dictionary::database db, service::migration_notifier& mn, service::migration_manager& mm, query_processor::memory_config mcfg, cql_config& cql_cfg)
         : _migration_subscriber{std::make_unique<migration_subscriber>(this)}
         , _proxy(proxy)
+        , _forwarder(forwarder)
         , _db(db)
         , _mnotifier(mn)
         , _mm(mm)