--- a/cql3/stats.hh
+++ b/cql3/stats.hh
@@ -76,6 +76,7 @@ struct cql_stats {
     int64_t select_allow_filtering = 0;
     int64_t select_partition_range_scan = 0;
     int64_t select_partition_range_scan_no_bypass_cache = 0;
+    int64_t select_parallelized = 0;
 
 private:
     uint64_t _unpaged_select_queries[(size_t)ks_selector::SIZE] = {0ul};