--- a/db/config.hh
+++ b/db/config.hh
@@ -339,6 +339,8 @@ public:
     named_value<bool> reversed_reads_auto_bypass_cache;
     named_value<bool> enable_optimized_reversed_reads;
     named_value<bool> enable_cql_config_updates;
+    named_value<bool> enable_parallelized_aggregation;
+
     named_value<uint16_t> alternator_port;
     named_value<uint16_t> alternator_https_port;
     named_value<sstring> alternator_address;