--- a/gms/feature.hh
+++ b/gms/feature.hh
@@ -139,6 +139,7 @@ extern const std::string_view SEPARATE_PAGE_SIZE_AND_SAFETY_LIMIT;
 extern const std::string_view SUPPORTS_RAFT_CLUSTER_MANAGEMENT;
 extern const std::string_view USES_RAFT_CLUSTER_MANAGEMENT;
 extern const std::string_view TOMBSTONE_GC_OPTIONS;
+extern const std::string_view PARALLELIZED_AGGREGATION;
 
 }