--- a/gms/feature_service.cc
+++ b/gms/feature_service.cc
@@ -64,6 +64,7 @@ constexpr std::string_view features::SEPARATE_PAGE_SIZE_AND_SAFETY_LIMIT = "SEPA
 constexpr std::string_view features::SUPPORTS_RAFT_CLUSTER_MANAGEMENT = "SUPPORTS_RAFT_CLUSTER_MANAGEMENT";
 constexpr std::string_view features::USES_RAFT_CLUSTER_MANAGEMENT = "USES_RAFT_CLUSTER_MANAGEMENT";
 constexpr std::string_view features::TOMBSTONE_GC_OPTIONS = "TOMBSTONE_GC_OPTIONS";
+constexpr std::string_view features::PARALLELIZED_AGGREGATION = "PARALLELIZED_AGGREGATION";
 
 static logging::logger logger("features");
 
@@ -93,6 +94,7 @@ feature_service::feature_service(feature_config cfg) : _config(cfg)
         , _supports_raft_cluster_mgmt(*this, features::SUPPORTS_RAFT_CLUSTER_MANAGEMENT)
         , _uses_raft_cluster_mgmt(*this, features::USES_RAFT_CLUSTER_MANAGEMENT)
         , _tombstone_gc_options(*this, features::TOMBSTONE_GC_OPTIONS)
+        , _parallelized_aggregation(*this, features::PARALLELIZED_AGGREGATION)
         , _raft_support_listener(_supports_raft_cluster_mgmt.when_enabled([this] {
             // When the cluster fully supports raft-based cluster management,
             // we can re-enable support for the second gossip feature to trigger
@@ -217,6 +219,7 @@ std::set<std::string_view> feature_service::known_feature_set() {
         gms::features::SUPPORTS_RAFT_CLUSTER_MANAGEMENT,
         gms::features::USES_RAFT_CLUSTER_MANAGEMENT,
         gms::features::TOMBSTONE_GC_OPTIONS,
+        gms::features::PARALLELIZED_AGGREGATION,
     };
 
     for (const sstring& s : _config._disabled_features) {
@@ -326,6 +329,7 @@ void feature_service::enable(const std::set<std::string_view>& list) {
         std::ref(_supports_raft_cluster_mgmt),
         std::ref(_uses_raft_cluster_mgmt),
         std::ref(_tombstone_gc_options),
+        std::ref(_parallelized_aggregation),
     })
     {
         if (list.contains(f.name())) {