--- a/gms/feature_service.hh
+++ b/gms/feature_service.hh
@@ -91,6 +91,7 @@ private:
     gms::feature _supports_raft_cluster_mgmt;
     gms::feature _uses_raft_cluster_mgmt;
     gms::feature _tombstone_gc_options;
+    gms::feature _parallelized_aggregation;
 
     gms::feature::listener_registration _raft_support_listener;
 
@@ -189,6 +190,10 @@ public:
 
     bool cluster_supports_tombstone_gc_options() const {
         return bool(_tombstone_gc_options);
+    }
+
+    bool cluster_supports_parallelized_aggregation() const {
+        return bool(_parallelized_aggregation);
     }
 
     static std::set<sstring> to_feature_set(sstring features_string);