--- a/idl/forward_request.idl.hh
+++ b/idl/forward_request.idl.hh
@@ -16,5 +16,5 @@ struct forward_result {
     std::vector<bytes_opt> query_results;
 };
 
-verb forward_request(query::forward_request) -> query::forward_result;
+verb forward_request(query::forward_request, std::optional<tracing::trace_info>) -> query::forward_result;
 }