… docs/design-notes/parallel_aggregations.md
… message/messaging_service.cc
… query.cc
… query_ranges_to_vnodes.cc
… query_ranges_to_vnodes.hh
… service/forward_service.cc
… service/forward_service.hh
… service/storage_proxy.cc
… service/storage_proxy.hh
… test/boost/cql_query_test.cc
… test/boost/storage_proxy_test.cc
… test/lib/cql_test_env.cc