@avikivity 2021-08-16
> 
> ## Detection
> To develop a mechanism for speeding up `count(*)` queries, there was a need to detect which queries have a `count(*)` selector. Due to this pull request being a proof of concept, detection was realized rather poorly. It is only allows catching the simplest cases of `count(*)` queries (with only one selector and no column name specified).


I've never seen a COUNT(*) query that would fail this detector. So this part will work just fine for production.