@avikivity 2021-08-16
Looks good. Please add metrics so we can count how many requests are sent and received per node in this new layer.