@piodul 2021-08-19
> > When a super-coordinator receives an `count(*)` query, it splits it into sub-queries. It does so, by splitting original query's partition ranges into list of vnodes, grouping them by their owner and creating sub-queries with partition ranges set to successive results of such grouping. After creating, each sub-query is sent to the owner of their partition ranges. Owner becomes a coordinator and executes received sub-query.
> 
> It is not clear to me, whether the coordinator receiving such a sub-query is expected to contact other nodes, or strictly execute it itself. In the latter case we are talking about query pushdown, not forwarding and I don't think we need a brand new mechanism for it (forwarding), we can extend the existing read verbs to optionally accept a `mapper`, which is applied on the results. Or maybe we need a new verb, but on the replica side it would just reuse the regular read path with the `mapper` applied on the results.

I think it is supposed to be the first case. Data from multiple replicas must be merged before the aggregate can be applied. This task can be divided by the "super-coordinator" into sub-ranges, then each sub-range delegated to other nodes/shards - "coordinators" - and finally the results can be merged by the "super-coordinator". This way we can utilize multiple CPU cores.