@denesb 2021-08-23
> I think it is supposed to be the first case. Data from multiple replicas must be merged before the aggregate can be applied. This task can be divided by the "super-coordinator" into sub-ranges, then each sub-range delegated to other nodes/shards - "coordinators" - and finally the results can be merged by the "super-coordinator". This way we can utilize multiple CPU cores.

I think we can do better. One of the reason aggregate queries are so slow is that all data in the table is sent to the coordinator, which then aggregates it. We should aim at eliminating this data movement altogether by splitting the range into sub-ranges such that each sub-range is owned by a single primary replica. This replica would then become the coordinator for this sub-range. It would read the range locally, requesting only digests from other replicas (to satisfy CL). In case of digest mismatch it would fall-back to the slow data merging method. This can be done as a follow-up.