@piodul 2021-08-23
> I think we can do better. One of the reason aggregate queries are so slow is that all data in the table is sent to the coordinator, which then aggregates it. We should aim at eliminating this data movement altogether by splitting the range into sub-ranges such that each sub-range is owned by a single primary replica. This replica would then become the coordinator for this sub-range. It would read the range locally, requesting only digests from other replicas (to satisfy CL). In case of digest mismatch it would fall-back to the slow data merging method. This can be done as a follow-up.

In this RFC, the queried range is divided with a per-vnode granularity. Each sub-range is sent to the first live replica of the corresponding vnode. See the [`storage_proxy::dispatch`](https://github.com/scylladb/scylla/pull/9209/commits/f60c9a5e5b3928edfc4a6d55b4527ada30594218#diff-b973f4e30060a6f4beaf81347658f565fa895998c849d92e80f574a6255bad75R4326-R4386) function.