@havaker 2021-11-07
v2:
- created a new forward_service to get rid of circular dependency between storage_proxy and cql3
- allocated new connection for forward_request handling
- rewritten some parts of the code to use coroutines
- rebased & applied other suggestions from the review process