@havaker 2021-11-19
v3:
- detecting `count(*)` at prepare time (new `select_statement` subclass)
- auto-generation of verb boilerplate
- using `column_id` instead of string to choose columns in `forward_request`
- removed `page_size` from `forward_request`
- added support for multiple results in `forward_result`
- added new source files to `CMakeLists.txt`
- fixed missing copyright/license blurb
- added tracepoints
- dropped `message/messaging_service.hh` include from `forward_service.hh`